Changed the way configs work
-config_orig.py is now a centralized config that all the other configs
import from and overwrite where necessary.

-Had to create two new classes CustomArgumentParser and CustomArgumentGroup
to add functionality that allows overwriting of already added argparse arguments

-If you don't see any argument in one of the other configs besides the original one
that you'd like to be different by default than the original one, then just add
it to the corresponding config file with a different default value. For example,
max_epoch is present in config_ours.py so that by default it runs for the correct number
of epochs and we don't have to specify that from the command line manually