Perplexity correlation analysis
-validation_set_performance_correlation.py looks at the correlations between the
difference ways of evaluating model perplexity (i.e. when it is compute using evaluate()
or derive() or get_perplexity_multibatch()). The point is to make sure that when using
just the first batch of the validation set to evaluate a model, this perplexity is
a good surrogate for the perplexity on the entire validation set