Changed calls to evaluate() to use entire validation set
-Now that we are using batch size of 64 regardless if we're deriving archiectures
or checking performance on test set, we are able to call evaluate() on entire validation
set without incurring significant computational penalities compared to when it was flat.

-The rewards that the controller gets when training is still just based on the first
part of the dataset of shape [35, 64] for computational reasons. But keep in mind this is
computing the perplexity on 64 distinct sequences of length 35, and NOT like computing the
perplexity on the first 35*64=2240 words of the dataset

-Fixed type in ablation studies that made it not load the trainer for randomly trained networks