Changed default log_step to 1
-The issue is that with a higher number, the best params might not end up getting
saved to file assuming that they occur somewhere in the middle of training and not
at the very end.

-Added saving of config for train_scratch to the file name "scratch_params.json".
So now in a given log directory, the "params.json" file is the config that was
used to trained ENAS, and the "scratch_params".json" file is the config that was
used to train a sampled architecture from scratch.